ingesta/mysql-rooms-api/requirements.txt
boto3==1.34.0
pandas==2.1.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0